More cleaning of getDistanceFor
